refactor: 🗂️ rename files to follow .routes and .controller naming convention
