+++ server/controllers/admin.controller.js
@@ -4,7 +4,7 @@ const User = require("../models/userModel");
 const Profile = require("../models/userProfileModel");
 const mailTemplateGenrator = require("../services/emailTemplateGenrator");
 const sendEmail = require("../services/sendEmailService");
-const { constants, tokens } = require("../utils/");
+const { constants, tokens } = require("../utils");
 const Status = require("../models/statusModel");
 const Partner = require("../models/partnerModel");
 const bcrypt = require("bcrypt");
