+++ server/controllers/user.controller.js
@@ -6,7 +6,7 @@ const jwt = require("jsonwebtoken");
 const asyncHandler = require("express-async-handler");
 const genratedOTP = require("../services/otpGenration");
 const mailTemplateGenrator = require("../services/emailTemplateGenrator");
-const { constants, templates, tokens } = require("../utils/");
+const { constants, templates, tokens } = require("../utils");
 const sendEmail = require("../services/sendEmailService");
 const Profile = require("../models/userProfileModel");
 const Transaction = require("../models/transactionsModel");
