+++ server/routes/api/user.routes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const userController = require("../../controllers/userController");
+const userController = require("../../controllers/user.controller");
 const validateToken = require("../../middleware/validateTokenHandler");
 const verifyGoogleIdToken = require("../../middleware/verifyGoogleIdToken");
